chore (refs T22614): rename all Tiptap libs files.
