+++ client/js/components/core/DpEditor/libs/editorBuildSuggestion.js
