+++ client/js/components/core/DpEditor/libs/editorCustomImage.js
