+++ client/js/components/core/DpEditor/libs/editorCustomInsert.js
@@ -10,19 +10,19 @@
 import { Mark } from 'tiptap'
 import { toggleMark } from 'tiptap-commands'
 
-export default class TiptapCustomMark extends Mark {
+export default class EditorCustomInsert extends Mark {
   get name () {
-    return 'mark'
+    return 'insert'
   }
 
   get schema () {
     return {
       parseDOM: [
         {
-          tag: 'mark'
+          tag: 'ins'
         }
       ],
-      toDOM: () => ['mark', { title: Translator.trans('text.mark') }, 0]
+      toDOM: () => ['ins', { title: Translator.trans('text.inserted') }, 0]
     }
   }
