+++ demosplan/DemosPlanCoreBundle/Resources/client/js/components/DpAnonymizeText.vue
@@ -56,9 +56,9 @@ import {
 import { Editor, EditorContent, EditorMenuBubble } from 'tiptap'
 import PreventDrop from '@DpJs/components/core/DpEditor/libs/preventDrop'
 import PreventKeyboardInput from '@DpJs/components/core/DpEditor/libs/preventKeyboardInput'
-import TiptapAnonymize from '@DpJs/components/core/DpEditor/libs/tiptapAnonymize'
-import TiptapObscure from '@DpJs/components/core/DpEditor/libs/tiptapObscure'
-import TiptapUnAnonymize from '@DpJs/components/core/DpEditor/libs/tiptapUnAnonymize'
+import EditorAnonymize from '@DpJs/components/core/DpEditor/libs/editorAnonymize'
+import EditorObscure from '@DpJs/components/core/DpEditor/libs/editorObscure'
+import EditorUnAnonymize from '@DpJs/components/core/DpEditor/libs/editorUnAnonymize'
 
 export default {
   name: 'DpAnonymizeText',
@@ -110,8 +110,9 @@ export default {
       disableInputRules: true,
       disablePasteRules: true,
       extensions: [
-        new TiptapAnonymize(),
-        new TiptapUnAnonymize(),
+        new EditorAnonymize(),
+        new EditorUnAnonymize(),
+        new EditorObscure(),
         new PreventKeyboardInput(),
         new PreventDrop(),
         new Bold(),
@@ -121,7 +122,6 @@ export default {
         new ListItem(),
         new Underline(),
         new History(),
-        new TiptapObscure(),
         new HardBreak()
       ],
       onUpdate: () => {
